# src/swagger.yaml
openapi: 3.0.1
info:
  title: CRUD AIP
  description: 'This is a Simple API Restful'
  version: 1.0.0
servers:
  - url: http://localhost:5000
...
paths:
...
  /user/refresh_token:
    get:
      tags:
        - user
      security:
        - ApiKeyAuth: []
      summary: Get refresh_token of user with access_token
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      iat:
                        type: integer
                      exp:
                        type: integer
                  access_token:
                    type: string
        400:
          description: Bad request
          content: {}
        403:
          description: Forbidden
          content: {}
        404:
          description: Not Found
          content: {}
        500:
          description: Internal server error
          content: {}
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
security:
  - cookieAuth: []